Split tests into setup.js and index.js (#342)
* first steps towards splitting tests

* finish splitting tests

* appease the linter

* strict mode and return

* remove extension
